Move operator tokenizing functions to op_tokenizing.c
