+++ src/tokens/op_tokenizing.c
@@ -6,11 +6,38 @@
 /*   By: gmoulin <gmoulin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/31 17:14:57 by gmoulin           #+#    #+#             */
-/*   Updated: 2025/01/30 15:32:38 by gmoulin          ###   ########.fr       */
+/*   Updated: 2025/01/30 16:33:20 by gmoulin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
+#include "tokens.h"
+#include "parsing.h"
+
+t_token	*tokenize_rdin(char **rl)
+{
+	if (is_space_tab((*rl)[1]) || !(*rl)[1])
+		return (new_token(REDIR_IN, ft_strdup(">")));
+	else if (is_rdin((*rl)[1]) && (is_space_tab((*rl)[2]) || !(*rl)[2]))
+		return (new_token(REDIR_APPEND, ft_strdup(">>")));
+	return (NULL);
+}
+
+t_token	*tokenize_rdout(char **rl)
+{
+	if (is_space_tab((*rl)[1]) || !(*rl)[1])
+		return (new_token(REDIR_OUT, ft_strdup("<")));
+	else if (is_rdout((*rl)[1]) && (is_space_tab((*rl)[2]) || !(*rl)[2]))
+		return (new_token(REDIR_HEREDOC, ft_strdup("<<")));
+	return (NULL);
+}
+
+t_token	*tokenize_pipe(char **rl)
+{
+	if (is_space_tab((*rl)[1]) || !(*rl)[1])
+		return (new_token(PIPE, ft_strdup("|")));
+	return (NULL);
+}
 
 int	op_tokenizing(char **rl, t_token **head)
 {
